refactor: expose native adapter discovery
