+++ src/winml/modelkit/sysinfo/__init__.py
@@ -9,6 +9,7 @@ from .device import (
     resolve_device,
     resolve_eps,
 )
+from .dxcore_adapters import DXCoreAdapterInfo, enumerate_compute_adapters
 from .hardware import CPU, GPU, NPU, get_available_devices
 from .luid import format_pdh_luid, get_ep_device_luid
 from .software import OS
@@ -17,10 +18,12 @@ from .sysinfo import SysInfo
 
 __all__ = [
     "CPU",
+    "DXCoreAdapterInfo",
     "GPU",
     "NPU",
     "OS",
     "SysInfo",
+    "enumerate_compute_adapters",
     "format_pdh_luid",
     "get_available_devices",
     "get_device_ep_map",
